Tweaking doc parsing and display
Updated styles.
Split methods.js into multiple files.
Removed subdoc processing.
