+++ common/main.jsx
@@ -12,6 +12,7 @@ const analytics = analytics || null;
 const AppRoutes = (
   <Route component={Layout} path="/">
     <Route component={Docs} path="/:repo/:branch/:alias" />
+    <Route component={Docs} path="/:repo/:branch/:alias/:subdoc" />
     <Route component={RedocAdmin} path="/redoc" />
     <IndexRoute component={Docs} />
   </Route>
