+++ package.json
@@ -26,12 +26,12 @@
     "highlight.js": "^9.0.0",
     "history": "^1.17.0",
     "jsdoc-parse": "^1.2.5",
-    "markdown-it": "^5.0.2",
+    "markdown-it": "^6.0.1",
     "markdown-it-replace-link": "^1.0.0",
     "punycode": "^1.4.0",
-    "react": "^0.14.7",
+    "react": "^0.14.8",
     "react-cookie": "^0.4.3",
-    "react-dom": "^0.14.7",
+    "react-dom": "^0.14.8",
     "react-helmet": "^2.3.1",
     "react-komposer": "^1.4.1",
     "react-lazy-load": "^2.0.2",
