+++ packages/redoc-base-theme/default/mixins.less
@@ -41,11 +41,11 @@ body.modal-open {
 }
 
 /* Loading indicator, spreads over whole screen estate to intercept mouse events until view is rendered. */
-.spinner-container {
-  left: 0;
-  position: absolute;
-  top: 0;
-}
+// .spinner-container {
+//   left: 0;
+//   position: absolute;
+//   top: 0;
+// }
 
 .editable-error-block {
     max-width: 100% !important;
