+++ packages/redoc-base-theme/styles/base.less
@@ -1,4 +1,8 @@
 
+html, body {
+  font-size: 16px;
+}
+
 #react-app {
   position: fixed;
   top: 0;
@@ -37,4 +41,27 @@
 
 .redoc.profile-image img {
   width: 100%;
+}
+
+.redoc.spinner-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100;
+}
+
+.redoc.spinner {
+  width: 50px;
+  height: 50px;
+  border-radius: 25px;
+  border-width: 3px;
+  border-style: solid;
+  border-color: @black80 @black80 transparent transparent;
+  animation: spinner 1.2s infinite linear;
+}
+
+@keyframes spinner {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
 }
