+++ packages/redoc-base-theme/styles/controls.less
@@ -36,7 +36,6 @@
   top: 0;
   right: 5px;
   z-index: 0;
-  font-size: 1.2rem;
 }
 
 .redoc.control .icon.left {
