+++ packages/redoc-base-theme/styles/header.less
@@ -24,7 +24,7 @@
 .redoc.header .title {
   display: flex;
   align-items: center;
-  font-size: 1.6rem;
+  font-size: 1.2rem;
   font-weight: 300;
   padding-left: 20px;
 }
