+++ packages/redoc-base-theme/styles/sidebar.less
@@ -21,20 +21,20 @@
 .guide-sub-subnav-item.level-1 {
   padding-left: 20px;
   font-size: 14px;
-  font-weight: 600;
+  font-weight: 400;
   // padding-top: 10px;
   text-transform: none;
 }
 
 .guide-sub-subnav-item.level-2 {
-  padding-left: 40px;
+  padding-left: 20px;
   font-size: 14px;
   text-transform: none;
   font-weight: 400;
 }
 
 .guide-sub-subnav-item.level-3 {
-  padding-left: 60px;
+  padding-left: 40px;
   font-size: 12px;
   font-weight: 400;
   text-transform: none;
@@ -100,7 +100,9 @@
 }
 
 .redoc.sidebar .menu .guide-nav-item {
-  font-size: 1.6rem;
+  margin-top: 1rem;
+  padding-top: 1rem;
+  border-top: 1px solid @border-color;
   margin-bottom: 10px;
   // margin-top: 50px;
   // text-transform: uppercase;
@@ -109,11 +111,21 @@
   &:hover {
     color: @black;
   }
+
+  &:first-child-of-type {
+    border-top: none;
+  }
+
+  a {
+    text-transform: uppercase;
+    color: @black;
+    font-weight: 600;
+  }
 }
 
 .redoc.sidebar .menu .reaction-nav-item {
   display: none;
-  font-size: 1.6rem;
+  font-size: 1rem;
   text-transform: uppercase;
   font-weight: 600;
   color: @black;
@@ -144,8 +156,13 @@
 
 .redoc.sidebar .menu .guide-sub-nav-item a {
   color: @black;
+  font-size: 0.9rem;
   &:hover {
     color: @black;
     text-decoration: underline;
   }
+}
+
+.redoc.sidebar .menu .guide-sub-nav-item.active > a {
+  color: @reaction-primary-color;
 }
