+++ packages/redoc-base-theme/styles/toolbar.less
@@ -3,7 +3,8 @@
   display: flex;
   justify-content: flex-end;
   align-items: center;
-  background-color: @black60;
+  background-color: @white;
+  border-bottom: 1px solid @border-color;
   height: 44px;
 }
 
@@ -13,7 +14,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  color: #fff;
+  // color: #fff;
   background-color: transparent;
   border: @black50;
   height: 100%;
