+++ packages/redoc-core/lib/plugins/api.js
@@ -52,6 +52,7 @@ function APIParser(md) {
     let isExampleSection = false;  // Startes with an H5 or H6, ends with H4 and above.
 
     state.tokens.forEach((token) => {
+      console.log(token);
       if (token.type === "heading_open") {
         const level = headingLevel(token.tag);
 
@@ -89,6 +90,27 @@ function APIParser(md) {
           isExampleSection = false;
         } else if (level >= 5 && inSection) {
           isExampleSection = true;
+        }
+      } else if (token.markup.indexOf("</article>") > -1) {
+        isExampleSection = false;
+
+        if (inSection === true) {
+          tokens = [
+            // Get the tokens as they are
+            ...tokens,
+
+            // Then concat the following
+            closeContentBlock(), // Close the content block of the section
+            openExampleBlock(), // Open an example block <aside>
+            ...sectionTokens, // Add any tokens to that block
+            closeExampleBlock(), // Close the example block </aside>
+            closeSection() // Close the section
+          ];
+
+          // then clean up section tokens
+          sectionTokens = [];
+
+          inSection = false;
         }
       }
